+++ mitmproxy/proxy/layers/http/_http1.py
@@ -116,9 +116,20 @@ class Http1Connection(HttpConnection, metaclass=abc.ABCMeta):
                 assert self.request
                 if h11_event.headers:
                     raise NotImplementedError(f"HTTP trailers are not implemented yet.")
-                if self.request.data.method.upper() != b"CONNECT":
-                    yield ReceiveHttp(self.ReceiveEndOfMessage(self.stream_id))
+
+                # Send EndOfMessage for all regular traffic, and for non-successful CONNECT responses (4xx, 5xx, etc), which may have a body.
+                # Skip EndOfMessage for CONNNECT requests (no request body) and successful CONNECT responses (to allow proper transition to passthrough).
                 is_request = isinstance(self, Http1Server)
+                is_connect = self.request.data.method.upper() == b"CONNECT"
+                is_connect_error_response = (
+                    not is_request
+                    and is_connect
+                    and self.response is not None
+                    and not (200 <= self.response.status_code < 300)
+                )
+
+                if not is_connect or is_connect_error_response:
+                    yield ReceiveHttp(self.ReceiveEndOfMessage(self.stream_id))
                 yield from self.mark_done(request=is_request, response=not is_request)
                 return